+++ docs/changelog.md
@@ -3,7 +3,9 @@
 # Version 3.3.5
 
 - *Effects* - Fixed `.screenSpaceAboveUI()` not correctly falling back to the below UI screenspace layer if the above UI layer is disabled
-- *Animations* - Fixed named sequence-crosshairs not working as a target position   
+- *Animations* - Fixed named sequence-crosshairs not working as a target position
+- *Sequencer* - Fixed minor Foundry v12 deprecation warnings
+- *Sequencer* - Updated PIXI filters
 
 # Version 3.3.4